Update development.md
An extended ramble about what community is and why it might be important. Feel free to reject if it is not what you had planned for this page, just felt like writing down some stuff that has been on my mind recently and this seemed like an appropriate place to do it.